+++ script/pre-compile.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 
 // relatived to root path of project
-const build = 'build';
+const build = 'out';
 if (fs.existsSync(build)) {
     fs.rmSync(build, {
         recursive: true,